pyproject.toml
[project]
name = "recruiting-challenge"
version = "0.1.0"
description = "FastAPI application that allows users to upload an image and generates a \"facial profile\" that describes key characteristics of the face which define its rea"
authors = [
    {name = "Neeha Rathna Janjanam",email = "samy1541998@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "insightface (>=0.7.3,<0.8.0)",
    "chromadb (>=1.0.13,<2.0.0)",
    "opencv-python",
    "streamlit (>=1.46.0,<2.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "onnxruntime (>=1.22.0,<2.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "torch (>=2.7.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

